refactor RtcConfigure into Create and Join
